Organize GraphService data classes into ShopifySharp.Services.Graph namespace
