+++ ShopifySharp.Tests/Services/Graph/GraphResultTests.cs
@@ -2,8 +2,8 @@
 using FakeItEasy;
 using FluentAssertions;
 using JetBrains.Annotations;
-using ShopifySharp.Graph;
 using ShopifySharp.Infrastructure.Serialization.Json;
+using ShopifySharp.Services.Graph;
 using Xunit;
 
 namespace ShopifySharp.Tests.Services.Graph;
