+++ ShopifySharp.Tests/Services/Graph/GraphService.ErrorHandlingTests.cs
@@ -3,8 +3,8 @@ using System.Threading.Tasks;
 using FakeItEasy;
 using FluentAssertions;
 using JetBrains.Annotations;
-using ShopifySharp.Graph;
 using ShopifySharp.Infrastructure.Serialization.Json;
+using ShopifySharp.Services.Graph;
 using Xunit;
 
 namespace ShopifySharp.Tests.Services.Graph;
