+++ ShopifySharp.Tests/Services/Graph/GraphService.PostAsyncTests.cs
@@ -10,9 +10,9 @@ using FakeItEasy.Configuration;
 using FluentAssertions;
 using JetBrains.Annotations;
 using Newtonsoft.Json.Linq;
-using ShopifySharp.Graph;
 using ShopifySharp.Infrastructure;
 using ShopifySharp.Infrastructure.Serialization.Json;
+using ShopifySharp.Services.Graph;
 using ShopifySharp.Tests.TestClasses;
 using ShopifySharp.Tests.Utilities;
 using Xunit;
