+++ ShopifySharp.Tests/Services/PartnerServiceTests.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
 using JetBrains.Annotations;
-using ShopifySharp.Infrastructure;
 using Xunit;
 
 namespace ShopifySharp.Tests.Services;
