+++ ShopifySharp.Tests/Services/ShopPlanServiceTests.cs
@@ -1,5 +1,3 @@
-using System.Net;
-using System.Net.Http;
 using System.Threading.Tasks;
 using FakeItEasy;
 using FluentAssertions;
