+++ ShopifySharp/Infrastructure/ShopifyGraphUserErrorsException.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ShopifySharp.Graph;
+using ShopifySharp.Services.Graph;
 
 namespace ShopifySharp;
