+++ ShopifySharp/Services/Graph/GraphRequestCostThrottleStatusExtension.cs
@@ -1,4 +1,4 @@
-namespace ShopifySharp;
+namespace ShopifySharp.Services.Graph;
 
 public record GraphRequestCostThrottleStatusExtension
 {
